Added Drag/Drop on EXE Functionality
I updated the code to address the request that was posted recently. You can now drag and drop files directly onto the EXE for slightly faster encrypt/decrypt in some cases. You must still select whether or not you are encrypting or decrypting the files and the standard GUI will popup, but instead of dragging and dropping, you can simply click. I don't think this really improves the program too much, but it was a relatively simple addition. I think context menus would be a better solution and I will be adding them in the next release.